Create version Convertors and clean up constructors
Also add class for weeds
