+++ src/main/java/com/agricraft/agricore/json/AgriJsonVersion.java
@@ -1,6 +1,7 @@
 package com.agricraft.agricore.json;
 
 import javax.annotation.Nullable;
+import java.util.function.Function;
 
 public interface AgriJsonVersion<T extends AgriSerializable> {
 
@@ -10,6 +11,11 @@ public interface AgriJsonVersion<T extends AgriSerializable> {
 
     @Nullable
     default AgriJsonVersion<?> previousVersion() {
+        return null;
+    }
+
+    @Nullable
+    default Function<AgriSerializable, T> versionConverter() {
         return null;
     }
 }
