--- src/main/java/com/agricraft/agricore/plant/old/v2/PlaceHolder.java
@@ -1,5 +1,0 @@
-package com.agricraft.agricore.plant.old.v2;
-
-public class PlaceHolder {
-    // For 1.16.4 which is the current active version
-}
