+++ src/main/java/com/agricraft/agricore/registry/AgriMutations.java
@@ -2,6 +2,8 @@ package com.agricraft.agricore.registry;
 
 import com.agricraft.agricore.json.AgriJsonVersion;
 import com.agricraft.agricore.plant.AgriMutation;
+import com.agricraft.agricore.plant.versions.v2.Versions_1_16;
+
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Collections;
@@ -10,18 +12,6 @@ import java.util.List;
 import java.util.Set;
 
 public class AgriMutations implements AgriLoadableRegistry<AgriMutation> {
-    public static final AgriJsonVersion<AgriMutation> VERSION = new AgriJsonVersion<AgriMutation>() {
-        @Override
-        public Class<AgriMutation> getElementClass() {
-            return AgriMutation.class;
-        }
-
-        @Override
-        public String descriptor() {
-            return "1.16.4";
-        }
-    };
-
     private final Set<AgriMutation> mutations;
 
     public AgriMutations() {
@@ -99,7 +89,7 @@ public class AgriMutations implements AgriLoadableRegistry<AgriMutation> {
 
     @Override
     public AgriJsonVersion<AgriMutation> getElementVersion() {
-        return VERSION;
+        return Versions_1_16.MUTATION;
     }
 
     @Override
